refactor kv-router services into brokerless namespace
