+++ lib/kv-router/src/services/indexer/listener.rs
@@ -11,7 +11,7 @@ use crate::protocols::{WorkerId, WorkerWithDpRank};
 use crate::recovery::{CursorObservation, CursorState};
 use crate::zmq_wire::{ZmqEventNormalizer, decode_event_batch};
 
-use super::indexer::Indexer;
+use super::backend::Indexer;
 use super::registry::ListenerRecord;
 use super::zmq::{
     MultipartMessage, SharedSocket, connect_dealer_socket, connect_sub_socket, recv_multipart,
@@ -541,7 +541,7 @@ async fn connect_replay_socket(
 mod tests {
     use super::{WATERMARK_UNSET, cursor_from_watermark};
     use crate::recovery::CursorObservation;
-    use crate::standalone_indexer::zmq::{
+    use crate::services::indexer::zmq::{
         SharedSocket, bind_pub_socket, connect_sub_socket, recv_multipart, send_multipart,
     };
