+++ lib/kv-router/src/services/indexer/metrics.rs
