+++ lib/kv-router/src/services/indexer/recovery.rs
