+++ lib/kv-router/src/services/slot_tracker/registry.rs
