--- lib/kv-router/src/standalone_indexer/runtime/mod.rs
@@ -1,6 +1,0 @@
-// SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-pub mod discovery;
-pub mod query_engine;
-pub mod subscriber;
